endpoint token control

+++ src/main/java/com/isssr/foodemperors/endpoint/DeliveryNoteEndPoint.java
@@ -2,15 +2,19 @@ package com.isssr.foodemperors.endpoint;
 
 import com.isssr.foodemperors.model.DeliveryNote;
 import com.isssr.foodemperors.service.DeliveryNoteService;
+import com.isssr.foodemperors.service.TokenService;
+import com.isssr.foodemperors.utils.TokenPayload;
 import org.springframework.web.bind.annotation.*;
 
 import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 /**
  * Created by Caim03 on 23/06/17.
  */
-
+//TODO verify users permissions
 @RestController
 @CrossOrigin(origins = "*")
 public class DeliveryNoteEndPoint {
@@ -18,28 +22,66 @@ public class DeliveryNoteEndPoint {
     @Inject
     private DeliveryNoteService deliveryNoteService;
 
+    @Inject
+    private TokenService tokenService;
+
     @RequestMapping(path = "api/delivery", method = RequestMethod.GET)
-    public List<DeliveryNote> searchAll() {
-        return deliveryNoteService.searchAll();
+    public List<DeliveryNote> searchAll(HttpServletRequest request, HttpServletResponse response) {
+        TokenPayload tokenPayload = tokenService.validateUser(request.getHeader("token"));
+        if (tokenPayload != null && (tokenPayload.getRole().equals("admin")
+                || tokenPayload.getRole().equals("manager") || tokenPayload.getRole().equals("wharehouseman")))
+            return deliveryNoteService.searchAll();
+        else{
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(path = "api/delivery", method = RequestMethod.POST)
-    public DeliveryNote saveDeliveryNote(@RequestBody DeliveryNote deliveryNote) {
-        return deliveryNoteService.saveDeliveryNote(deliveryNote);
+    public DeliveryNote saveDeliveryNote(@RequestBody DeliveryNote deliveryNote, HttpServletRequest request, HttpServletResponse response) {
+        TokenPayload tokenPayload = tokenService.validateUser(request.getHeader("token"));
+        if (tokenPayload != null && (tokenPayload.getRole().equals("admin")
+                || tokenPayload.getRole().equals("manager") || tokenPayload.getRole().equals("wharehouseman")))
+            return deliveryNoteService.saveDeliveryNote(deliveryNote);
+        else{
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(path = "api/delivery", method = RequestMethod.PUT)
-    public DeliveryNote updateDeliveryNote(@RequestBody DeliveryNote deliveryNote) {
-        return deliveryNoteService.updateDeliveryNote(deliveryNote);
+    public DeliveryNote updateDeliveryNote(@RequestBody DeliveryNote deliveryNote, HttpServletRequest request, HttpServletResponse response) {
+        TokenPayload tokenPayload = tokenService.validateUser(request.getHeader("token"));
+        if (tokenPayload != null && (tokenPayload.getRole().equals("admin")
+                || tokenPayload.getRole().equals("manager") || tokenPayload.getRole().equals("wharehouseman")))
+            return deliveryNoteService.updateDeliveryNote(deliveryNote);
+        else{
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(path = "api/delivery", method = RequestMethod.DELETE)
-    public DeliveryNote deleteDeliveryNote(@RequestBody DeliveryNote deliveryNote) {
-        return deliveryNoteService.deleteDeliveryNote(deliveryNote);
+    public DeliveryNote deleteDeliveryNote(@RequestBody DeliveryNote deliveryNote, HttpServletRequest request, HttpServletResponse response) {
+        TokenPayload tokenPayload = tokenService.validateUser(request.getHeader("token"));
+        if (tokenPayload != null && (tokenPayload.getRole().equals("admin")
+                || tokenPayload.getRole().equals("manager") || tokenPayload.getRole().equals("wharehouseman")))
+            return deliveryNoteService.deleteDeliveryNote(deliveryNote);
+        else{
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(path = "api/delivery/findBy/id/{id}", method = RequestMethod.GET)
-    public DeliveryNote searchById(@RequestParam String id) {
-        return deliveryNoteService.searchById(id);
+    public DeliveryNote searchById(@RequestParam String id, HttpServletRequest request, HttpServletResponse response) {
+        TokenPayload tokenPayload = tokenService.validateUser(request.getHeader("token"));
+        if (tokenPayload != null && (tokenPayload.getRole().equals("admin")
+                || tokenPayload.getRole().equals("manager") || tokenPayload.getRole().equals("wharehouseman")))
+            return deliveryNoteService.searchById(id);
+        else{
+            response.setStatus(401);
+            return null;
+        }
     }
 }
